revoltpy/requirements.txt
revolt.py
requests